refactor: common util to save api records
