+++ integrated_channels/utils.py
@@ -476,6 +476,18 @@ def get_enterprise_customer_model():
     """
     return apps.get_model('enterprise', 'EnterpriseCustomer')
 
+def cornerstone_request_log_model():
+    """
+    Returns the ``CornerstoneAPIRequestLogs`` class.
+    """
+    return apps.get_model("cornerstone", "CornerstoneAPIRequestLogs")
+
+
+def integrated_channel_request_log_model():
+    """
+    Returns the ``IntegratedChannelAPIRequestLogs`` class.
+    """
+    return apps.get_model("integrated_channel", "IntegratedChannelAPIRequestLogs")
 
 def get_enterprise_customer_from_enterprise_enrollment(enrollment_id):
     """
@@ -501,3 +513,53 @@ def get_enterprise_client_by_channel_code(channel_code):
         'canvas': CanvasAPIClient,
     }
     return _enterprise_client_model_by_channel_code[channel_code]
+
+def store_api_call(
+    enterprise_customer,
+    enterprise_customer_configuration_id,
+    endpoint,
+    payload,
+    time_taken,
+    status_code,
+    response_body,
+    channel_code="",
+    user_agent=None,
+    user_ip=None,
+):
+    """
+    Creates new record in CornerstoneAPIRequestLogs table.
+    """
+    try:
+        if channel_code == "CSOD":
+            cornerstone_request_log_model().objects.create(
+                user_agent=user_agent,
+                user_ip=user_ip,
+                enterprise_customer=enterprise_customer,
+                enterprise_customer_configuration_id=enterprise_customer_configuration_id,
+                endpoint=endpoint,
+                payload=payload,
+                time_taken=time_taken,
+                status_code=status_code,
+                response_body=response_body,
+            )
+        else:
+            integrated_channel_request_log_model().objects.create(
+                enterprise_customer=enterprise_customer,
+                enterprise_customer_configuration_id=enterprise_customer_configuration_id,
+                endpoint=endpoint,
+                payload=payload,
+                time_taken=time_taken,
+                status_code=status_code,
+                response_body=response_body,
+            )
+    except Exception as e:  # pylint: disable=broad-except
+        LOGGER.error(
+            f"[{channel_code}]: store_api_call raised error while storing API call: {e}"
+            f"user_agent={user_agent}, user_ip={user_ip}, enterprise_customer={enterprise_customer}"
+            f"enterprise_customer_configuration_id={enterprise_customer_configuration_id},"
+            f"endpoint={endpoint}"
+            f"payload={payload}"
+            f"time_taken={time_taken}"
+            f"status_code={status_code}"
+            f"response_body={response_body}"
+        )
